updating in Posts Edit

+++ routes/posts.js
@@ -39,20 +39,24 @@ router.put("/:id", async (req, res) => {
 router.delete("/:id", async (req, res) => {
     try {
         const post = await Post.findById(req.params.id);
-        if (post.userId === req.body.userId) {
-            await post.deleteOne({ $set: req.body });
-            res.status(200).json("The Post Has Been deleted");
-
+        if (post) {
+            if (post.userId === req.body.userId) {
+                await post.deleteOne({ $set: req.body });
+                res.status(200).json("The Post Has Been deleted");
+            }
+            else {
+                res.status(403).json("you Can Only delete your Own Post")
+            }
         }
         else {
-            res.status(403).json("you Can Only delete your Own Post")
+            res.status(500).json("Post Has been Deleted");
+
         }
     }
 
     catch (err) {
         res.status(500).json(err)
     }
-
 })
